refactor: extract theme into ThemeWrapper HOC component
